feat(jarvis): fixed

+++ gui/src/liveStatFetch.cpp
@@ -8,7 +8,7 @@ double getLiveCurrent() {
     int min = 1000;
     int max = 1500;
 
-    int randomValue = min + rand() % (max - min + 1);
+    double randomValue = min + rand() % (max - min + 1);
     randomValue = randomValue/1000;
     return randomValue;
 }
